refactor(chatbot): reorganize messaging services and flatten handler packages
- Relocate messaging-related services (`UiTextService`, `ResponseFactory`, `TextRenderer`, and `EmojiProvider`) to the `services.messages` package
- Remove `handlers` subpackages and move handler implementations directly into their respective parent feature packages (`core`, `submission`, and `verification`)
- Update imports across WhatsApp and Telegram integrations, notification services, and tests to reflect the restructured packages
