+++ chatbot/src/main/java/pt/estga/chatbot/features/core/MainMenuHandler.java
@@ -1,17 +1,15 @@
-package pt.estga.chatbot.features.core.handlers;
+package pt.estga.chatbot.features.core;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import pt.estga.chatbot.constants.CallbackData;
 import pt.estga.chatbot.constants.MessageKey;
 import pt.estga.chatbot.context.*;
-import pt.estga.chatbot.features.core.MainMenuFactory;
 import pt.estga.chatbot.models.BotInput;
 import pt.estga.chatbot.models.BotResponse;
-import pt.estga.chatbot.services.UiTextService;
+import pt.estga.chatbot.services.messages.UiTextService;
 
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 
 @Component
