+++ chatbot/src/main/java/pt/estga/chatbot/features/submission/SubmissionStartHandler.java
@@ -1,12 +1,11 @@
-package pt.estga.chatbot.features.core.handlers;
+package pt.estga.chatbot.features.submission;
 
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import pt.estga.chatbot.context.ChatbotContext;
 import pt.estga.chatbot.context.ConversationState;
 import pt.estga.chatbot.context.ConversationStateHandler;
-import pt.estga.chatbot.context.CoreState;
 import pt.estga.chatbot.context.HandlerOutcome;
+import pt.estga.chatbot.context.SubmissionState;
 import pt.estga.chatbot.models.BotInput;
 import pt.estga.chatbot.models.BotResponse;
 
@@ -14,26 +13,26 @@ import java.util.Collections;
 import java.util.List;
 
 @Component
-@RequiredArgsConstructor
-public class StartHandler implements ConversationStateHandler {
+public class SubmissionStartHandler implements ConversationStateHandler {
 
     @Override
     public HandlerOutcome handle(ChatbotContext context, BotInput input) {
-        context.setCurrentState(CoreState.MAIN_MENU);
-        return new HandlerOutcome.Success();
+        return new HandlerOutcome.Redispatch();
     }
 
     @Override
     public ConversationState canHandle() {
-        return CoreState.START;
+        return SubmissionState.SUBMISSION_STATE;
+    }
+
+    @Override
+    public boolean isAutomatic() {
+        return true;
     }
 
     @Override
     public ConversationState getNextState(ChatbotContext context, ConversationState currentState, HandlerOutcome outcome, BotInput input) {
-        if (outcome instanceof HandlerOutcome.Failure) {
-            return currentState;
-        }
-        return CoreState.MAIN_MENU;
+        return SubmissionState.WAITING_FOR_PHOTO;
     }
 
     @Override
