+++ chatbot/src/main/java/pt/estga/chatbot/services/messages/UiTextService.java
@@ -1,4 +1,4 @@
-package pt.estga.chatbot.services;
+package pt.estga.chatbot.services.messages;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.MessageSource;
