+++ chatbot/src/main/java/pt/estga/chatbot/telegram/services/TelegramEmojiProvider.java
@@ -2,7 +2,7 @@ package pt.estga.chatbot.telegram.services;
 
 import org.springframework.stereotype.Component;
 import pt.estga.chatbot.constants.EmojiKey;
-import pt.estga.chatbot.services.EmojiProvider;
+import pt.estga.chatbot.services.messages.EmojiProvider;
 
 @Component("telegramEmojiProvider")
 public class TelegramEmojiProvider implements EmojiProvider {
