+++ chatbot/src/test/java/pt/estga/chatbot/features/submission/SubmissionFlowStrategyTest.java
@@ -7,10 +7,9 @@ import pt.estga.chatbot.context.CoreState;
 import pt.estga.chatbot.context.HandlerOutcome;
 import pt.estga.chatbot.context.SubmissionState;
 import pt.estga.chatbot.features.core.MainMenuFactory;
-import pt.estga.chatbot.features.submission.handlers.*;
 import pt.estga.chatbot.models.BotInput;
 import pt.estga.fileapi.FileStorageOperations;
-import pt.estga.chatbot.services.UiTextService;
+import pt.estga.chatbot.services.messages.UiTextService;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.mock;
