+++ chatbot/src/test/java/pt/estga/chatbot/services/ConversationDispatcherTest.java
@@ -16,6 +16,7 @@ import pt.estga.chatbot.models.BotInput;
 import pt.estga.chatbot.models.BotResponse;
 import pt.estga.chatbot.models.Platform;
 import pt.estga.chatbot.models.text.RichText;
+import pt.estga.chatbot.services.messages.UiTextService;
 
 import java.util.List;
